BASE-2281: cachestore_rediscluster: RedisCluster specific cache store and session handler
The current implementation of the cache store and session handler for redis
cluster are built on top of the existing redis store and session handler - each
of those then also duplicates the code used to talk to the redis servers
themselves.

This plugin reduces the duplication of code between these two components whilst
also being streamlined to focus on working with just RedisCluster rather than
both Redis and RedisCluster.

Lastly, it includes improvements to cover edge cases the original versions do
not. Including, but not limited to:

* race condition when getting a lock and setting its expiry - its possible,
  though unlikely, for a session to obtain a lock and then fail to set the
  expiry time out on that lock. This is fixed in this plugin by setting the
  expiry time in the same command the lock is obtained in.

* built-in support for various purge methods, including future proofing with
  support for the UNLINK command in Redis 4.

* generate a uniqid from php for the key used when purging data in
  PURGEMODE_LAZY. This reduces the calls to redis needed during a purge
  operation from 3 to 2.

Functionally, end users will see no change from using this plugin.

Change-Id: I39338f5e6605af4f5b3f076df9b327f4880f5989
Reviewed-on: https://gerrit.netspot.com.au/25521
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim@netspot.com.au>