BASE-2312: cachestore_rediscluster: Ratelimiting and locking support
Adds a class with some helper functions to make it possible to
check if a particular action has exceeded a rate limit.

Currently supports two limits:
* count: Has there been more than N accesses of the resource in
  the time window.
* average: Does the average of the value for the accesses in the
  time window exceed a particular value.

Locking:

Adds a class to provide simple application level locking. Whilst the
cachestore API provides locking, it doesn't provide the extra features
provided here, namely the ability to have a TTL on the lock for the rare
event a thread with a lock dies without releasing it.

Change-Id: I18c390faee98d6ea653b42be0d290826ea5e8c1e
Reviewed-on: https://gerrit.netspot.com.au/25822
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim@netspot.com.au>