BASE-2320: cachestore_rediscluster: Set correct session expiry time
When setting the expiry time for the session lock, redis needs to be told how
many seconds until the key should expire. The plugin was mistakenly telling it
the absolute time since epoch, which results in a much larger expiry than
expected.

This change corrects that behaviour, and also addresses a potential expiry
issue when tracking how many requests a session has waiting. It would increment
the number and set the expiry for that in separate calls - these now occur in a
single transaction.

Change-Id: I44efc9afa69b7d6556d8276941419b3b42a8eb27
Reviewed-on: https://gerrit.netspot.com.au/25886
Tested-by: Jenkins
Reviewed-by: Mat Gebhardt <mat.gebhardt@netspot.com.au>