BASE-2609: cachestore_rediscluster: Add support for UNLINK command
Redis4 introduced a new command - UNLINK - which allows you to delete data from
redis in a non-blocking fashion (it does it in the background).

However, the php extension for redis (phpredis) does not yet support the
command directly. But it can still be used by calling it via phpredis'
rawCommand.

Change-Id: I3bbba8dca48e526c1059f6b00bd9b1c53b3a4cd4
Reviewed-on: https://gerrit.netspot.com.au/28347
Reviewed-by: Tim Lock <tim@netspot.com.au>
Tested-by: Jenkins