BASE-2847: cachestore_rediscluster: make test checks more resilient to fast runs
Without this change, its possible for a system running fast enough to
cause the ratelimit checking to fail.

It's deemed acceptable for actual use of the ratelimiter to allow an
extra request through the system before kicking in given a microsecond
is miniscule compared to the usual window we're rate limiting over
(2 minutes).

Change-Id: Iee42e5d77995a4e1ade612b1b1c2c0f86cfe6b9d
Reviewed-on: https://gerrit.netspot.com.au/29699
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim@netspot.com.au>