BASE-3333: cachestore_rediscluster: Support config controlled site identifier
Support a siteidentifier set in CFG, and when not set, use one composed
from just the dbname and dbprefix without md5sum'ing it.

We previously summed it due to a maxlength in memcache key names - which
are not a concern with redis. Keeping them in plain text makes it far
easier to identify which keys belong to which service when digging
through redis manually.

Change-Id: Iba0d287242eb200e81061c82906e8460264334f2
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/34125
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim@netspot.com.au>