BASE-3626: cachestore_rediscluster: Fine tune the redis lock polling
The rediscluster session handler polls for a lock, and it waits a random time
from 100-1000ms so on average each blocked http request waits 550ms.

This change is a port of MDL-68577 for rediscluster and reduces the average
wait time from 550ms to 110ms.

After 5 seconds, the delay is increased to 1 second.

Change-Id: I9edcb364a654646941eb9a3fb18efac402585bac
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/35610
Reviewed-by: Tim Lock <tim.lock@openlms.net>
Tested-by: Jenkins