BASE-3843: cachestore_rediscluster: Add support for compression
phpredis 5.1.0 added the ability to configure compression for data stored in redis.

This change adds an option to the RedisCluster cache store to configure a
compression type. It adds support for both LZF and ZSTD compression algorithms
(LZF if you have phpredis 4.0.0+, ZSTD if you have phpredis 5.1.0+ and LZ4 if
you have 5.3.0+).

This also removes the 'none' option for serialization - its not a supported
method for the data MUC stores.

Change-Id: I7401f82dcb1ad041c3df321e84997deb3ff80c5d
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/37234
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim.lock@openlms.net>