BASE-4042: cachestore_rediscluster: Support Session mode definitions
This allows us to map session-mode definitions against rediscluster out of the
box. Without it the UI doesn't let us choose it.

This also tidies up the code to meet coding style standards and most mess
detector ones (some are the fault of the Caching API and can't really be met).

Change-Id: I155c3f4156cab80683c63c027caeb803e166ddf7
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/38083
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim.lock@openlms.net>