BASE-4372: cachestore_rediscluster: When searching use hscan over hkeys and support shards
Without this - find_all and find_by_prefix calls use hkeys to list the keys in
a definition. For large caches - this is a long blocking operation. It also
doesn't work on sharded caches prior to this change.

So here we switch to using hScan to iterate over the definitions hash (or
hashes when sharding is involved).

The find_by_prefix approach is also improved by passing the prefix we care
about to redis as part of the hScan call.

While we're here, this also fixes failures in the session unit tests.

Change-Id: I7aba0291bb1ad6b3a7db631d86a92f5d4409504a
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/39890
Reviewed-by: Tony Levi <tony.levi@openlms.net>
Tested-by: Jenkins