BASE-4574: cachestore_rediscluster: Add waiting-room functionality to session-waiter
Without this, we have plenty of users with long running requests holding extra
php threads open waiting on a session lock. This is bad for our systems
(threads being wasted) and bad for users (they see a blank screen for up to a
minute).

This change introduces the ability to show the user a "please wait" type page
when the session lock is taking awhile to become available.

This functionality is only available for GET requests to non-AJAX scripts.

Change-Id: I670e929ffefd6c13fd5d1bbff401891cc3e9a595
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/40956
Reviewed-by: Tony Levi <tony.levi@openlms.net>
Tested-by: Jenkins